libpod: Move functions related to checkpoints to container_internal_common.go
[NO NEW TESTS NEEDED]
